(refactor) modal to make things a little simpler
