--- src/components/common/ModalSubTitle/index.tsx
@@ -1,19 +1,0 @@
-import styled from 'styled-components'
-
-export const ModalSubTitle = styled.h3`
-  color: ${(props) => props.theme.colors.darkGrey};
-  font-size: 16px;
-  font-weight: normal;
-  font-stretch: normal;
-  font-style: italic;
-  line-height: 1.13;
-  letter-spacing: normal;
-  text-align: left;
-  margin: 0;
-  overflow: hidden;
-  padding: 0;
-  position: relative;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  z-index: 1;
-`
